+++ uzywasie/CHANGELOG.md
@@ -4,8 +4,14 @@ All notable changes to the "uzywasie" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [0.0.4] - 2025-02-08
+## [1.0.6] - 2025-02-08
+
+### Added
+
+- Added command `extension.jsonlGetMessage`
+
+## [0.0.4 - 0.0.5] - 2025-02-08
 
 ### Changed
 
-- Remove unused command `onCommand:extension.oddPipeReplace`
+- Remove unused command `extension.oddPipeReplace`